refactor: refactor PageTitle import path
