+++ src/components/ui/PageTitle/PageTitle.tsx
@@ -1,8 +1,9 @@
-import { useRouter } from "next/router";
 import { FC, useCallback } from "react";
+import { useRouter } from "next/router";
+import cn from "clsx";
+
 import Breadcrumb, { BreadcrumbProps } from "../Breadcrumb";
 import { PrimaryButton } from "../Buttons";
-import cn from "clsx";
 
 export type PageTitleProps = {
   title: string;
